[CLEANUP] Restructure internal-test-helpers exports
